refactor: moved equality checks to specific util classes (and tests)
