+++ frontend/tests/src/util/QuUtilsTest.spec.ts
@@ -1,6 +1,7 @@
 import { describe, expect, it } from "vitest";
 
 import { useQu } from "@/util/QuUtils";
+import Himmelsrichtung from "@/types/enum/Himmelsrichtung";
 
 describe("QuUtils", () => {
   it("computeAvailableKnotenarme: entfernt Duplikate und sortiert aufsteigend", () => {
@@ -34,4 +35,59 @@ describe("QuUtils", () => {
     const result = computeAvailableKnotenarme(zaehlung);
     expect(result).toEqual([]);
   });
+
+  describe("areVerkehrsbeziehungenEqual", () => {
+    const { areQuerungsverkehreEqual } = useQu();
+    it("returns true for two identical arrays", () => {
+      const a = [
+        { knotenarm: 1, richtung: Himmelsrichtung.N },
+        { knotenarm: 2, richtung: Himmelsrichtung.O },
+      ] as any;
+      const b = [
+        { knotenarm: 1, richtung: Himmelsrichtung.N },
+        { knotenarm: 2, richtung: Himmelsrichtung.O },
+      ] as any;
+      expect(areQuerungsverkehreEqual(a, b)).toBe(true);
+    });
+
+    it("is order independent", () => {
+      const a = [
+        { knotenarm: 1, richtung: Himmelsrichtung.N },
+        { knotenarm: 2, richtung: Himmelsrichtung.O },
+      ] as any;
+      const b = [
+        { knotenarm: 2, richtung: Himmelsrichtung.O },
+        { knotenarm: 1, richtung: Himmelsrichtung.N },
+      ] as any;
+      expect(areQuerungsverkehreEqual(a, b)).toBe(true);
+    });
+
+    it("treats undefined as empty array", () => {
+      expect(areQuerungsverkehreEqual(undefined, [])).toBe(true);
+      expect(areQuerungsverkehreEqual([], undefined)).toBe(true);
+      expect(areQuerungsverkehreEqual(undefined, undefined)).toBe(true);
+    });
+
+    it("returns false for different lengths", () => {
+      const a = [{ knotenarm: 1, richtung: Himmelsrichtung.N }] as any;
+      const b = [
+        { knotenarm: 1, richtung: Himmelsrichtung.N },
+        { knotenarm: 2, richtung: Himmelsrichtung.O },
+      ] as any;
+      expect(areQuerungsverkehreEqual(a, b)).toBe(false);
+    });
+
+    it("returns false when richtung differs", () => {
+      const a = [{ knotenarm: 1, richtung: Himmelsrichtung.N }] as any;
+      const c = [{ knotenarm: 1, richtung: Himmelsrichtung.O }] as any; // different richtung
+      expect(areQuerungsverkehreEqual(a, c)).toBe(false);
+    });
+
+    it("returns false when knotenarm differs", () => {
+      const a = [{ knotenarm: 1, richtung: Himmelsrichtung.N }] as any;
+      const d = [{ knotenarm: 2, richtung: Himmelsrichtung.N }] as any; // different knotenarm
+      expect(areQuerungsverkehreEqual(a, d)).toBe(false);
+    });
+  });
+
 });
